backend/requirements.txt
asgiref==3.6.0
botocore==1.29.59
certifi==2022.12.7
charset-normalizer==3.0.1
cloudinary==1.31.0
dj-database-url==1.2.0
Django==4.1.5
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.14.0
environs==9.5.0
factory-boy==3.2.1
Faker==16.6.1
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
marshmallow==3.19.0
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.3.0
